day7 - Very simple stupid solution
I'm not proud of this. It's just bruteforce. My first idea was to make
an algorithm that guesses how the possible results will be sorted
depending on the operator placements. This could have made possible to
use binary search on the operator placements. However, while writing it
I looked at the input and said "meh, it's not that big, I can probably
bruteforce it in a reasonable time, and write the code very fast". I
did not write the code extremely fast, but I guessed right that it would
have worked, even if it's O(Lp^(N-1)), where L is the number of rows in
the input, N is the average number of numbers to be considered for the
result, and p is 2 in first part and 3 in the second one. Maybe I will
write the other version I had in mind or look up if someone else has
done something similar