day 20
Today it was hard. I had an idea and it worked. I took a while though to
realize that it could work just as well for part 2. But at least now
it's clearer what the idea is: I can use the fact that the path is a
single possible path, so after calculating it, I can use the difference
in index to understand which is the distance through the path between
two points on it. In order to calculate the new distance after the
cheat, I can instead just consider manhattan distance. So in part 2 I
considered all pairs of indexes (i1, i2) which lead to cells of
manhattan distance "d" \le then 20, and then just (i2 - i1) - d is the
saved distance