add day23 v2 + myGraphs
I didn't want to use any non-standard library in AoC, so here it comes
the update where I remove the need for using Graphs.jl. In order to not
modify that heavily the old solution, I made a very simple
UnorderedGraph type in file myGraphs.jl and implemented the exact
functions I needed in the file. The idea is that this might eventually
become my own Graphs library, to be used in next editions or AoC or just
for fun. Actually, I have written the code in it in one evening and I
didn't have Internet connections, so I guessed an algo for cliques,
which I don't actually know. So I coded a terrible way of finding them
but it turns out that it works, so for now it's fine. In the future I
might decide to add also BFS, DFS, minimal spanning tree, and expand
the "library" to consider also directed and weighted graphs, hence
adding also Dijkstra and A*. In other words, it might become a fun
playground for revising Graph algorithms.

Finally, this is probably the last AoC 2024 commit for the next months